requirements.txt
opencv-python
numpy
face-recognition
Pillow
pygame
ultralytics
matplotlib